+++ src/presenter/movie-presenter.js
@@ -1,7 +1,7 @@
 import PopupView from '../view/popup-view';
 import CardsView from '../view/cards-view.js';
 import {renderElement, remove} from '../mock/render';
-import {RenderPosition} from '../mock/generate.js';
+import {RenderPosition} from '../const.js';
 import {replace} from '../mock/utils/utils';
 
 export const PopupMode = {
