+++ src/view/footer-view.js
@@ -1,12 +1,15 @@
 import AbstractView from './abstract-view.js';
-import {POSTSCOUNT} from './extra-view.js';
 
 const createFooterTemplate = () => (
-  `<p>${POSTSCOUNT} movies inside</p>`
+  '<p> 0 movies inside</p>'
 );
 
 export default class FooterView extends AbstractView {
   get template() {
     return createFooterTemplate();
+  }
+
+  setCount(count) {
+    this.element.textContent = `${count} movies inside`;
   }
 }
